+++ packages/ui/src/GenericInputForm/AbiFunctionName.tsx
@@ -4,9 +4,8 @@ import { AbiFunction } from "viem";
 import { useCallback } from "react";
 import { FunctionSignature } from "./FunctionSignature";
 import { TbAlertCircle } from "react-icons/tb";
-import { generateInitialValues } from "./utils";
+import { resetAbiFunctionParams } from "./utils";
 import { AbiInputParam } from "./types";
-import { isArray } from "ramda-adjunct";
 
 export const AbiFunctionName = () => {
     const form = useFormContext();
@@ -30,24 +29,10 @@ export const AbiFunctionName = () => {
             ).find((abiFunction) => abiFunction.name === abiFunctionName);
 
             if (nextAbiFunction) {
-                const emptyFunctionParams: AbiInputParam[] = [];
-                (nextAbiFunction.inputs as AbiInputParam[]).forEach((input) => {
-                    generateInitialValues(input, emptyFunctionParams);
-                });
-
-                const prevAbiFunctionParams =
-                    form.getInputProps("abiFunctionParams");
-
-                if (isArray(prevAbiFunctionParams.value)) {
-                    prevAbiFunctionParams.value.forEach((_, index) => {
-                        form.setFieldError(
-                            `abiFunctionParams.${index}.value`,
-                            null,
-                        );
-                    });
-                }
-
-                form.setFieldValue("abiFunctionParams", emptyFunctionParams);
+                resetAbiFunctionParams(
+                    form,
+                    nextAbiFunction.inputs as AbiInputParam[],
+                );
             }
         },
         [combobox, form, selectedSpecification],
@@ -76,6 +61,7 @@ export const AbiFunctionName = () => {
                                         {...form.getInputProps(
                                             "abiFunctionName",
                                         )}
+                                        style={{ overflow: "hidden" }}
                                         onClick={() =>
                                             combobox.toggleDropdown()
                                         }
